rpc: serve writable event before the readable one.
In case the RPC client performs a request with a timeout on the
response, it is more reliable to run PROTOBUF_C_RPC_EVENT_WRITABLE first
and have the response sent promptly.

After the PROTOBUF_C_RPC_EVENT_WRITABLE is completed, there is plenty
of time for a PROTOBUF_C_RPC_EVENT_READABLE to complete even if it
sleeps or block.

Signed-off-by: Emmanuel Roullit <emmanuel.roullit@gmail.com>